other way to find user id

+++ main.py
@@ -1,5 +1,7 @@
 import asyncio
 import logging
+import types
+
 from aiogram import Bot, Dispatcher
 from aiogram.types import Message
 from config import Config, load_config
@@ -34,24 +36,26 @@ async def delete_message_with_url(message: Message):
     warning_message = await message.answer(f"@{username}, сообщения с гиперссылками запрещены.")
     await asyncio.create_task(delete_after_delay(warning_message, 30))
 
-@dp.message(is_read_only)
+@dp.message_handler(content_types=types.ContentType.ANY)
 async def delete_message_from_read_only(message: Message):
-    text = message.text[:30]
+    user_id = message.from_user.id
+    if user_id in config.read_only:
+        text = message.text[:30]
 
-    try:
-        await message.delete()
-        log_message = f"Удалили сообщение от пользователя {message.from_user.username} ({message.from_user.id}), который писал: {text}"
-        logging.info(log_message)
-        await send_log_to_admin(log_message)
+        try:
+            await message.delete()
+            log_message = f"Удалили сообщение от пользователя {message.from_user.username} ({message.from_user.id}), который писал: {text}"
+            logging.info(log_message)
+            await send_log_to_admin(log_message)
 
-        warning_message = await message.answer(
-            f"@{message.from_user.username}, вам запрещено писать сообщения в этом чате.")
-        await asyncio.create_task(delete_after_delay(warning_message, 10))
+            warning_message = await message.answer(
+                f"@{message.from_user.username}, вам запрещено писать сообщения в этом чате.")
+            await asyncio.create_task(delete_after_delay(warning_message, 10))
 
-    except Exception as e:
-        error_message = f"❌ Не удалось удалить сообщение от {message.from_user.username} в чате {message.chat.id}: {e}"
-        logging.error(error_message)
-        await send_log_to_admin(error_message)
+        except Exception as e:
+            error_message = f"❌ Не удалось удалить сообщение от {message.from_user.username} в чате {message.chat.id}: {e}"
+            logging.error(error_message)
+            await send_log_to_admin(error_message)
 
 async def delete_after_delay(message: Message, delay: int):
     await asyncio.sleep(delay)
